faster?

+++ color.h
@@ -4,7 +4,7 @@
 #include "vec3.h"
 #include <iostream>
 
-void write_color(std::ostream &out, const color& pixel_color, int samples_per_pixel){
+void write_color(std::ostream &out, const color& pixel_color, int64_t samples_per_pixel){
     //write the translated [0,255] value of each color component
     auto r = pixel_color.x();
     auto g = pixel_color.y();
@@ -17,9 +17,9 @@ void write_color(std::ostream &out, const color& pixel_color, int samples_per_pi
     b = sqrt(scale * b);
 
     // Write the translated [0,255] value of each color component.
-    out << static_cast<int>(256 * clamp(r, 0.0, 0.999)) << ' '
-        << static_cast<int>(256 * clamp(g, 0.0, 0.999)) << ' '
-        << static_cast<int>(256 * clamp(b, 0.0, 0.999)) << '\n';
+    out << static_cast<int64_t>(256 * clamp(r, 0.0, 0.999)) << ' '
+        << static_cast<int64_t>(256 * clamp(g, 0.0, 0.999)) << ' '
+        << static_cast<int64_t>(256 * clamp(b, 0.0, 0.999)) << '\n';
 }
 
 #endif
