+++ rtweekend.h
@@ -14,12 +14,12 @@ using std::sqrt;
 
 // Constants
 
-const float infinity = std::numeric_limits<float>::infinity();
-const float pi = 3.1415926535897932385;
+const double infinity = std::numeric_limits<double>::infinity();
+const double pi = 3.1415926535897932385;
 
 // Utility Functions
 
-inline float degrees_to_radians(float degrees) {
+inline double degrees_to_radians(double degrees) {
     return degrees * pi / 180.0;
 }
 
@@ -27,17 +27,17 @@ inline float degrees_to_radians(float degrees) {
 
 // Utility
 
-inline float random_double() {
+inline double random_double() {
     // Returns a random real in [0,1).
     return rand() / (RAND_MAX + 1.0);
 }
 
-inline float random_double(float min, float max) {
+inline double random_double(double min, double max) {
     // Returns a random real in [min,max).
     return min + (max-min)*random_double();
 }
 
-inline float clamp(float x, float min, float max) {
+inline double clamp(double x, double min, double max) {
     if (x < min) return min;
     if (x > max) return max;
     return x;
